[Builtins] Restructure everything
