+++ plutus-core/docs/Builtins.md
@@ -322,7 +322,7 @@ liftBool b = if b then truePlc else falsePlc
 
 where `truePlc` and `falsePlc` are the ASTs for `\x y -> x` and `\x y -> y` respectively.
 
-Writing such a conversion function is not a big deal and we can easily generalize it, after all Haskell types/values are convertible to Plutus Core types/terms by design. Problems arise when you need to go in the other direction: unlift a Plutus Core term into a Haskell value. Such unlifting is a much more complicated beast. We did it previously ([docs available here](https://github.com/effectfully/plutus-prototype/blob/7a071a5ed2276ae3119f48fa0ca307422ed86ec2/language-plutus-core/src/Language/PlutusCore/Constant/Typed.hs#L201-L252)), but eventually we decided that every built-in function must reference only those types that are in the universe. This way lifting always amounts to wrapping a constant with a constructor like `ConstantInteger` and unlifting always amounts to directly unwrapping from such a constructor.
+Writing such a conversion function is not a big deal and we can easily generalize it, after all Haskell types/values are convertible to Plutus Core types/terms by design. Problems arise when you need to go in the other direction: unlift a Plutus Core term into a Haskell value. Such unlifting is a much more complicated beast. We did it previously ([docs available here](https://github.com/effectfully/plutus-prototype/blob/7a071a5ed2276ae3119f48fa0ca307422ed86ec2/language-plutus-core/src/Language/PlutusCore.Builtin/Typed.hs#L201-L252)), but eventually we decided that every built-in function must reference only those types that are in the universe. This way lifting always amounts to wrapping a constant with a constructor like `ConstantInteger` and unlifting always amounts to directly unwrapping from such a constructor.
 
 ## Saturated vs unsaturated built-in functions
