+++ plutus-core/generators/PlutusCore/Generators/Interesting.hs
@@ -21,7 +21,7 @@ module PlutusCore.Generators.Interesting
     , fromInterestingTermGens
     ) where
 
-import PlutusCore.Constant
+import PlutusCore.Builtin
 import PlutusCore.Core
 import PlutusCore.Default
 import PlutusCore.Evaluation.Result
