+++ plutus-core/generators/PlutusCore/Generators/Internal/Entity.hs
@@ -32,7 +32,7 @@ import PlutusCore.Generators.Internal.TypeEvalCheck
 import PlutusCore.Generators.Internal.TypedBuiltinGen
 import PlutusCore.Generators.Internal.Utils
 
-import PlutusCore.Constant
+import PlutusCore.Builtin
 import PlutusCore.Core
 import PlutusCore.Default
 import PlutusCore.Evaluation.Result
