+++ plutus-core/plutus-core.cabal
@@ -46,9 +46,9 @@ library
         PlutusCore.Check.Scoping
         PlutusCore.Check.Uniques
         PlutusCore.Check.Value
-        PlutusCore.Constant
-        PlutusCore.Constant.Debug
-        PlutusCore.Constant.Dynamic.Emit
+        PlutusCore.Builtin
+        PlutusCore.Builtin.Debug
+        PlutusCore.Builtin.Emitter
         PlutusCore.Core
         PlutusCore.Data
         PlutusCore.DataFilePaths
@@ -175,10 +175,13 @@ library
         common
     other-modules:
         PlutusCore.Analysis.Definitions
-        PlutusCore.Constant.Function
-        PlutusCore.Constant.Kinded
-        PlutusCore.Constant.Meaning
-        PlutusCore.Constant.Typed
+        PlutusCore.Builtin.HasConstant
+        PlutusCore.Builtin.KnownKind
+        PlutusCore.Builtin.KnownType
+        PlutusCore.Builtin.KnownTypeAst
+        PlutusCore.Builtin.Meaning
+        PlutusCore.Builtin.Polymorphism
+        PlutusCore.Builtin.TypeScheme
         PlutusCore.Core.Instance
         PlutusCore.Core.Instance.Eq
         PlutusCore.Core.Instance.Pretty
