+++ plutus-core/plutus-core/src/PlutusCore/Builtin.hs
@@ -1,0 +1,14 @@
+-- | Reexports from modules from the @Constant@ folder.
+
+module PlutusCore.Builtin
+    ( module Export
+    ) where
+
+import PlutusCore.Builtin.Emitter as Export
+import PlutusCore.Builtin.HasConstant as Export
+import PlutusCore.Builtin.KnownKind as Export
+import PlutusCore.Builtin.KnownType as Export
+import PlutusCore.Builtin.KnownTypeAst as Export
+import PlutusCore.Builtin.Meaning as Export
+import PlutusCore.Builtin.Polymorphism as Export
+import PlutusCore.Builtin.TypeScheme as Export
