--- plutus-core/plutus-core/src/PlutusCore/Constant.hs
@@ -1,11 +1,0 @@
--- | Reexports from modules from the @Constant@ folder.
-
-module PlutusCore.Constant
-    ( module Export
-    ) where
-
-import PlutusCore.Constant.Dynamic.Emit as Export
-import PlutusCore.Constant.Function as Export
-import PlutusCore.Constant.Kinded as Export
-import PlutusCore.Constant.Meaning as Export
-import PlutusCore.Constant.Typed as Export
