+++ plutus-core/plutus-ir/src/PlutusIR/Analysis/Dependencies.hs
@@ -7,7 +7,7 @@
 module PlutusIR.Analysis.Dependencies (Node (..), DepGraph, StrictnessMap, runTermDeps, runTypeDeps) where
 
 import PlutusCore qualified as PLC
-import PlutusCore.Constant qualified as PLC
+import PlutusCore.Builtin qualified as PLC
 import PlutusCore.Name qualified as PLC
 
 import PlutusIR
