+++ plutus-core/untyped-plutus-core/test/Evaluation/Builtins/Common.hs
@@ -10,7 +10,7 @@ module Evaluation.Builtins.Common
     ) where
 
 import PlutusCore qualified as TPLC
-import PlutusCore.Constant
+import PlutusCore.Builtin
 import PlutusCore.Default
 import PlutusCore.Evaluation.Machine.ExMemory
 import PlutusCore.Evaluation.Machine.MachineParameters
