+++ plutus-core/untyped-plutus-core/test/Evaluation/Builtins/MakeRead.hs
@@ -6,7 +6,7 @@ module Evaluation.Builtins.MakeRead
     ) where
 
 import PlutusCore qualified as TPLC
-import PlutusCore.Constant
+import PlutusCore.Builtin
 import PlutusCore.Default
 import PlutusCore.Evaluation.Machine.Exception
 import PlutusCore.Evaluation.Result
